ajouter des resources a des taches et fixer se droit de update et delete par rapport a un projet

+++ back/app/Http/Controllers/ResourceController.php
@@ -25,11 +25,19 @@ class ResourceController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+
+    public function store(Request $request,$project)
     {
-        $resource = new Resource($request[0]);
-        $resource->save();
-        return $resource; 
+        request()->validate([
+            'name' => 'required',
+        ]);
+        $data = $request->all();
+        $data['project_id'] = $project;
+        $ressource = Resource::create($data);
+        return response()->json([
+            'status' => 'success',
+            'data' => $ressource
+        ]);
     }
 
     /**
@@ -38,6 +46,7 @@ class ResourceController extends Controller
      * @param  \App\Resource  $resource
      * @return \Illuminate\Http\Response
      */
+
     public function show($id)
     {
         return new ResourcesResource(Resource::find($id));
@@ -50,10 +59,27 @@ class ResourceController extends Controller
      * @param  \App\Resource  $resource
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+
+    public function update(Request $request, $project,$id)
     {
-        DB::table('resources')->where('id', $id)->update($request[0]);
-        return response()->json([$request[0], 200]);
+        $token = $request->header('Authorization');
+        $token = substr($token, 6);
+        $token = trim($token);
+        $user = getUserInfo($token);
+        $user_id = $user->id;
+        $checkRight = checkRight($user_id,$project);
+        if($checkRight){
+            $resource = Resource::find($id);
+            $resource->name = $request->name;
+            $resource->ratio = $request->ratio;
+            $resource->job = $request->job;
+            $resource->first_name = $request->first_name;
+            $resource->save();
+            InsertLog("UpdateRessource",$id,$user_id);
+            return response()->json(["message"=> "updated"],200);
+        }else{
+            return response()->json(["message"=> "Unauthorized delete"],401);
+        }
     }
 
     /**
@@ -62,9 +88,26 @@ class ResourceController extends Controller
      * @param  \App\Resource  $resource
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+
+    public function destroy(Request $request,$project,$id)
     {
-        DB::table('resources')->where('id', $id)->delete();
-        return response()->json(null, 204);
+        $token = $request->header('Authorization');
+        $token = substr($token, 6);
+        $token = trim($token);
+        $user = getUserInfo($token);
+        $user_id = $user->id;
+        $checkRight = checkRight($user_id,$project);
+        if($checkRight){
+            DB::table('resources')->where('id', $id)->delete();
+            $data =  DB::table('tasks')
+                ->where('resource_id', $id)
+                ->update(['resource_id' => Null]);
+            InsertLog("deleteRessource",$id,$user_id);
+            return response()->json(['message' => 'ressource delete'], 200);
+        }else{
+            return response()->json(['message' => 'Unauthorized delete'], 401);
+
+        }
+
     }
 }
